+++ launchpy/configs.py
@@ -1,7 +1,7 @@
 import json, os
 from pathlib import Path
 from typing import Optional
-from .config import config_object, header
+from launchpy.config import config_object, header
 
 def find_path(path: str) -> Optional[Path]:
     """Checks if the file denoted by the specified `path` exists and returns the Path object
@@ -30,7 +30,7 @@ def createConfigFile(filename:str='config_launch',auth_type: str = "oauthV2", sc
             scope="https://ims-na1.adobelogin.com/s/ent_reactor_admin_sdk"
             or 
             scope="https://ims-na1.adobelogin.com/s/ent_reactor_sdk"
-        auth_type : OPTIONAL : The type of Oauth type you want to use for your config file. Possible value: "jwt" or "oauthV2"
+        auth_type : OPTIONAL : The type of Oauth type you want to use for your config file. Possible value: "oauthV2"
     """
     json_data = {
         'org_id': '<orgID>',
@@ -39,10 +39,6 @@ def createConfigFile(filename:str='config_launch',auth_type: str = "oauthV2", sc
     }
     if auth_type == 'oauthV2':
         json_data['scopes'] = "<scopes>"
-    elif auth_type == 'jwt':
-        json_data["tech_id"] = "<something>@techacct.adobe.com"
-        json_data["pathToKey"] = "<path/to/your/privatekey.key>"
-        json_data['scope'] = scope
     if '.json' not in filename:
         filename = f"{filename}.json"
     with open(filename, 'w') as cf:
@@ -84,8 +80,6 @@ def importConfigFile(path: str = None,auth_type:str=None) -> None:
         if auth_type is None:
             if 'scopes' in provided_keys:
                 auth_type = 'oauthV2'
-            elif 'tech_id' in provided_keys and "pathToKey" in provided_keys:
-                auth_type = 'jwt'
         args = {
             "org_id" : provided_config['org_id'],
             "secret" : provided_config['secret'],
@@ -94,9 +88,6 @@ def importConfigFile(path: str = None,auth_type:str=None) -> None:
         }
         if auth_type == 'oauthV2':
             args["scopes"] = provided_config["scopes"].replace(' ','')
-        if auth_type == 'jwt':
-            args["tech_id"] = provided_config["tech_id"]
-            args["path_to_key"] = provided_config["pathToKey"]
         configure(**args)
 
 
@@ -129,8 +120,6 @@ def configure(org_id: str = None,
               tech_id: str = None,
               secret: str = None,
               client_id: str = None,
-              path_to_key: str=None,
-              private_key: str = None,
               scopes : str= None,
               scope: str="https://ims-na1.adobelogin.com/s/ent_reactor_admin_sdk",
               ):
@@ -140,8 +129,6 @@ def configure(org_id: str = None,
         tech_id : REQUIRED : Technical Account ID
         secret : REQUIRED : secret generated for your connection
         client_id : REQUIRED : The client_id (old api_key) provided by the JWT connection. 
-        path_to_key : REQUIRED : If you have a file containing your private key value.
-        private_key : REQUIRED : If you do not use a file but pass a variable directly.
         scope : OPTIONAL : Scope that is needed for JWT auth.
             Possible scope: https://www.adobe.io/authentication/auth-methods.html#!AdobeDocs/adobeio-auth/master/JWT/Scopes.md
     """
@@ -161,25 +148,8 @@ def configure(org_id: str = None,
     header["x-api-key"] = client_id
     config_object["tech_id"] = tech_id
     config_object["secret"] = secret
-    config_object["pathToKey"] = path_to_key
-    config_object["private_key"] = private_key
     config_object["official_scope"] = scope
     config_object["scopes"] = scopes
     # ensure the reset of the state by overwriting possible values from previous import.
     config_object["date_limit"] = 0
     config_object["token"] = ""
-
-
-def get_private_key_from_config(config: dict) -> str:
-    """
-    Returns the private key directly or read a file to return the private key.
-    """
-    private_key = config.get('private_key')
-    if private_key is not None:
-        return private_key
-    private_key_path = find_path(config['pathToKey'])
-    if private_key_path is None:
-        raise FileNotFoundError(f'Unable to find the private key under path `{config["pathToKey"]}`.')
-    with open(Path(private_key_path), 'r') as f:
-        private_key = f.read()
-    return private_key